change login/logout button position
