+++ frontend/app/diagrams/page.tsx
@@ -21,6 +21,7 @@ import {
 } from "@/components/ui/card";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Badge } from "@/components/ui/badge";
+import { SignInOutButton } from "@/components/sign-in-out-button";
 
 interface DiagramInfo {
   url: string;
@@ -80,14 +81,17 @@ export default function DiagramsPage() {
     <div className="min-h-screen bg-slate-50">
       <div className="container mx-auto p-6 max-w-7xl">
         {/* Header */}
-        <div className="mb-8">
-          <h1 className="text-3xl font-bold text-slate-900 mb-2">
-            Diagram Visualization
-          </h1>
-          <p className="text-slate-600">
-            Retrieve and visualize biomodel diagrams and network
-            representations.
-          </p>
+        <div className="mb-8 flex items-center justify-between gap-4">
+          <div>
+            <h1 className="text-3xl font-bold text-slate-900 mb-2">
+              Diagram Visualization
+            </h1>
+            <p className="text-slate-600">
+              Retrieve and visualize biomodel diagrams and network
+              representations.
+            </p>
+          </div>
+          <SignInOutButton />
         </div>
 
         {/* Input Form */}
