+++ frontend/app/search/page.tsx
@@ -35,6 +35,7 @@ import {
 } from "@/components/ui/collapsible";
 import Link from "next/link";
 import { cn } from "@/lib/utils";
+import { SignInOutButton } from "@/components/sign-in-out-button";
 
 interface SearchFilters {
   bmId: string;
@@ -156,19 +157,22 @@ export default function BiomodelSearchPage() {
     <div className="min-h-screen bg-slate-50">
       <div className="container mx-auto p-6 max-w-7xl">
         {/* Header */}
-        <div className="mb-8 flex items-center gap-4">
-          <div className="w-12 h-12 bg-blue-100 rounded-full flex items-center justify-center flex-shrink-0">
-            <Search className="h-6 w-6 text-blue-600" />
+        <div className="mb-8 flex items-center justify-between gap-4">
+          <div className="flex items-center gap-4">
+            <div className="w-12 h-12 bg-blue-100 rounded-full flex items-center justify-center flex-shrink-0">
+              <Search className="h-6 w-6 text-blue-600" />
+            </div>
+            <div>
+              <h1 className="text-3xl font-bold text-slate-900">
+                Biomodel Database Search
+              </h1>
+              <p className="text-slate-600 mt-1">
+                Search and explore biomodels from the VCell database with advanced
+                filtering options.
+              </p>
+            </div>
           </div>
-          <div>
-            <h1 className="text-3xl font-bold text-slate-900">
-              Biomodel Database Search
-            </h1>
-            <p className="text-slate-600 mt-1">
-              Search and explore biomodels from the VCell database with advanced
-              filtering options.
-            </p>
-          </div>
+          <SignInOutButton />
         </div>
 
         {/* Search Form */}
