+++ frontend/components/file-fetcher.tsx
@@ -12,6 +12,7 @@ import {
   CardTitle,
 } from "@/components/ui/card";
 import { Alert, AlertDescription } from "@/components/ui/alert";
+import { SignInOutButton } from "@/components/sign-in-out-button";
 import XMLViewer from "react-xml-viewer";
 
 export interface FileFetcherProps {
@@ -98,9 +99,12 @@ export default function FileFetcher({
     <div className="min-h-screen bg-slate-50">
       <div className="container mx-auto p-6 max-w-7xl">
         {/* Header */}
-        <div className="mb-8">
-          <h1 className="text-3xl font-bold text-slate-900 mb-2">{title}</h1>
-          <p className="text-slate-600">{description}</p>
+        <div className="mb-8 flex items-center justify-between gap-4">
+          <div>
+            <h1 className="text-3xl font-bold text-slate-900 mb-2">{title}</h1>
+            <p className="text-slate-600">{description}</p>
+          </div>
+          <SignInOutButton />
         </div>
         {/* Input Form */}
         <Card className="mb-8 shadow-sm border-slate-200">
